Mem-Change int to byte

+++ rain_setup.h
@@ -2,5 +2,5 @@
 #define RAIN_A_PIN A1
 
 int rain_val_A;
-int rain_val_D;
+byte rain_val_D;
 char rain_str[4];
